requirements.txt
requests==2.26.0
reportlab==3.6.5
googletrans==4.0.0-rc1
arabic-reshaper==3.1.0
python-bidi==0.4.2
fpdf==1.7.2
